+++ lib/LoggerUtils.ts
@@ -8,29 +8,25 @@ import {
   IOSetting,
   IOStd,
   IOReturnError,
-  IOErrorStack,
   IOPrefixOption,
   IOReturnType,
 } from "./LoggerInterfaces.js";
 import { hostname } from "os";
 import { format } from "util";
-import ErrorStackParser from "error-stack-parser";
-import { StackFrame } from "error-stack-parser";
 import type { ForegroundColor } from "chalk";
 import { LoggerEvent } from "./LoggerEvents.js";
-
-const parse = ErrorStackParser.parse.bind(ErrorStackParser);
+import { Helper } from "./HelperFunctions.js";
 
 export class LoggerUtils<P extends {}> extends LoggerEvent {
-  protected name: string = "";
-  protected cagetoryName: string = "";
-  protected hostname = hostname();
-  protected format: "hidden" | "json" | "pretty" = "hidden";
-  protected short: boolean = false;
-  protected childProps: P;
-  protected useColor: boolean;
-  private pid = process.pid;
-  protected levelLog: IOLevelLog = IOLevelLog.NONE;
+  protected _name: string = "";
+  protected _cagetoryName: string = "";
+  protected _hostname = hostname();
+  protected _format: "hidden" | "json" | "pretty" = "hidden";
+  protected _short: boolean = false;
+  protected _childProps: P;
+  protected _useColor: boolean;
+  private _pid = process.pid;
+  protected _levelLog: IOLevelLog = IOLevelLog.NONE;
   /**
    * This Is My Logger
    */
@@ -46,82 +42,31 @@ export class LoggerUtils<P extends {}> extends LoggerEvent {
     childOpt: P = {} as P
   ) {
     super();
-    this.name = instanceName;
-    this.cagetoryName = cagetoryName;
-    this.format = format;
-    this.short = short;
-    this.levelLog = levelLog;
-    this.childProps = childOpt;
-    this.useColor = useColor;
+    this._name = instanceName;
+    this._cagetoryName = cagetoryName;
+    this._format = format;
+    this._short = short;
+    this._levelLog = levelLog;
+    this._childProps = childOpt;
+    this._useColor = useColor;
     this.setMaxListeners(0);
   }
-  private formatString(...data: any[]): string {
+  private _formatString(...data: any[]): string {
     return format.apply(null, [...data]);
   }
-  private write(...data: any[]): void {
-    process.stdout.write(this.formatString(...data, "\n"));
+  private _write(...data: any[]): void {
+    process.stdout.write(this._formatString(...data, "\n"));
   }
   public getBindingOpt(): Readonly<P> {
-    return Object.freeze(this.childProps);
-  }
-  protected cleanPath(path: string | undefined): string {
-    if (path === undefined) return "";
-    return path.replace(/file:\/\/\//, "").replace(/%20/g, " ");
-  }
-  public getTimeAndType(
-    type: "Log" | "Error" | "Info" | "Warn" | "Fatal" | "Debug",
-    color: typeof ForegroundColor = (chalk.Color = "white"),
-    loggedAt: string
-  ): IOReturnGetTimeAndType {
-    const frame = parse(new Error("Genesis Error"))[3];
-    const filePath: string = this.cleanPath(frame.getFileName());
-    const lineNumber: number | undefined = frame.getLineNumber();
-    const lineColumm: number | undefined = frame.getColumnNumber();
-    const prefixString = this.useColor
-      ? chalk[color](
-          `[Type: ${type}, Time: ${loggedAt}, File: "${filePath}:${lineNumber}:${lineColumm}", PID: ${
-            this.pid
-          }] ${chalk.whiteBright(`[${chalk.cyanBright(this.cagetoryName)}]`)}`
-        )
-      : `[Type: ${type}, Time: ${loggedAt}, File: "${filePath}:${lineNumber}:${lineColumm}", PID: ${this.pid}] [${this.cagetoryName}]`;
-    return {
-      ToString: prefixString,
-    };
-  }
-  public getErrorStack(
-    stack?: StackFrame[] | undefined,
-    range: number = 3
-  ): IOErrorStack {
-    if (stack) {
-      return {
-        filePath: this.cleanPath(stack[0].getFileName()),
-        fullFilePath: stack[0].getFileName(),
-        lineNumber: stack[0].getLineNumber(),
-        lineColumm: stack[0].getColumnNumber(),
-        methodName: stack[0].getFunctionName(),
-        functionName: stack[0].getFunctionName(),
-        isConstructor: stack[0].getIsConstructor(),
-      };
-    } else {
-      const localStack = parse(new Error("Genesis Error"));
-      return {
-        filePath: this.cleanPath(localStack[range].getFileName()),
-        fullFilePath: localStack[range].getFileName(),
-        lineNumber: localStack[range].getLineNumber(),
-        lineColumm: localStack[range].getColumnNumber(),
-        methodName: localStack[range].getFunctionName(),
-        functionName: localStack[range].getFunctionName(),
-        isConstructor: localStack[range].getIsConstructor(),
-      };
-    }
+    return Object.freeze(this._childProps);
   }
   public setSettings({
-    instanceName = this.name,
-    cagetoryName = this.cagetoryName,
-    format = this.format,
-    levelLog = this.levelLog,
-    short = this.short,
-    useColor = this.useColor,
+    instanceName = this._name,
+    cagetoryName = this._cagetoryName,
+    format = this._format,
+    levelLog = this._levelLog,
+    short = this._short,
+    useColor = this._useColor,
   }: IOLoggerInterface): void {
     this.emit("settingChange", this.listSetting(), {
       ...this.listSetting(),
@@ -129,32 +74,32 @@ export class LoggerUtils<P extends {}> extends LoggerEvent {
       cagetoryName,
       format,
       levelLog,
-      hostName: this.hostname,
+      hostName: this._hostname,
       short,
       useColor,
     });
-    this.name = instanceName;
-    this.cagetoryName = cagetoryName;
-    this.format = format;
-    this.short = short;
-    this.useColor = useColor;
-    if (this.levelLog !== levelLog) {
-      this.emit("levelChange", levelLog, this.levelLog);
-      this.levelLog = levelLog;
+    this._name = instanceName;
+    this._cagetoryName = cagetoryName;
+    this._format = format;
+    this._short = short;
+    this._useColor = useColor;
+    if (this._levelLog !== levelLog) {
+      this.emit("levelChange", levelLog, this._levelLog);
+      this._levelLog = levelLog;
     }
   }
   public listSetting(): IOSetting {
     return {
-      instanceName: this.name,
-      cagetoryName: this.cagetoryName,
-      hostName: this.hostname,
-      format: this.format,
-      levelLog: this.levelLog,
-      useColor: this.useColor,
+      instanceName: this._name,
+      cagetoryName: this._cagetoryName,
+      hostName: this._hostname,
+      format: this._format,
+      levelLog: this._levelLog,
+      useColor: this._useColor,
     };
   }
   public get loggerName(): string {
-    return this.name;
+    return this._name;
   }
   public set loggerName(newName: string) {
     if (newName.length > 1) {
@@ -168,24 +113,13 @@ export class LoggerUtils<P extends {}> extends LoggerEvent {
       throw Error("newName error");
     }
   }
-  protected getLoggedTime(): string {
+  protected _getLoggedTime(): string {
     return `${
       new Date().toLocaleTimeString() + " " + new Date().toLocaleDateString()
     }`;
   }
-  public toJson(
-    data: any,
-    replacer?: ((this: any, key: string, value: any) => any) | undefined,
-    spacing?: number | undefined | string
-  ): string {
-    return JSON.stringify(data, replacer, spacing);
-  }
-  public toPretty<T extends any[]>(data: string): IOReturnType<T, P> {
-    return JSON.parse(data);
-  }
-  protected censor(censor: any) {
+  protected _censor(censor: any) {
     var i = 0;
-
     return function (key: string, value: any) {
       if (
         i !== 0 &&
@@ -194,80 +128,158 @@ export class LoggerUtils<P extends {}> extends LoggerEvent {
         censor == value
       )
         return "[Circular]";
-
       ++i;
-
       return value;
     };
   }
-  protected checkLevel(range: IOLevelLog): boolean {
-    return this.levelLog >= range;
+  protected _printFatalLog(
+    logObject: IOReturnType<IOReturnError[], P>,
+    errors: Error[] = []
+  ) {
+    let stringToPrint = "";
+    if (!this._short) {
+      if (!logObject.data) {
+      } else {
+        stringToPrint += this._formatString(
+          `${logObject.fullPrefix.ToString}\n`
+        );
+      }
+    } else {
+    }
+    stringToPrint +=
+      "--------------------------------------------------------------------------";
+    if (logObject.data.length < 1) return;
+    if (this._format === "pretty") {
+      this._write(stringToPrint);
+      for (var i = 0, len = logObject.data.length; i < len; i++) {
+        let err = logObject.data[i].defaultError;
+        this._writePrettyFatal(err);
+      }
+      this.emit(
+        "fatalLogging",
+        this._levelLog,
+        logObject,
+        new Date(),
+        logObject.fullPrefix.ToString,
+        errors
+      );
+    }
   }
-  protected handleLog<T extends any[]>(
+  protected _printPrettyLog<T extends any[]>(
+    logObject: IOReturnType<T, P>,
+    errors: Error[] = []
+  ): void {
+    let stringToPrint = "";
+    if (!this._short) {
+      if (!logObject.data) {
+      } else {
+        stringToPrint += this._formatString(
+          `${logObject.fullPrefix.ToString} `
+        );
+      }
+    } else {
+    }
+    stringToPrint += this._formatString(...logObject.data);
+    if (logObject.levelLog !== "fatal") {
+      this._write(stringToPrint);
+    } else {
+      this._printFatalLog(logObject, errors);
+    }
+  }
+  protected _printJsonLog<T extends any[]>(
+    logObject: IOReturnType<T, P>,
+    errors: Error[] = []
+  ): void {
+    if (logObject.levelLog === "fatal") {
+      this.emit(
+        "fatalLogging",
+        this._levelLog,
+        logObject,
+        new Date(),
+        logObject.fullPrefix.ToString,
+        errors
+      );
+    }
+    let stringToPrint = "";
+    if (!this._short) {
+      if (!logObject.data) {
+      } else {
+        stringToPrint += this._formatString(
+          `${logObject.fullPrefix.ToString} `
+        );
+      }
+    } else {
+    }
+    stringToPrint += this._formatString(
+      Helper.toJson.apply(this, [logObject, this._censor(logObject), 2])
+    );
+    this._write(stringToPrint);
+  }
+  protected _printHiddenLog<T extends any[]>(
+    logObject: IOReturnType<T, P>,
+    errors: Error[] = []
+  ) {
+    if (logObject.levelLog === "fatal") {
+      this.emit(
+        "fatalLogging",
+        this._levelLog,
+        logObject,
+        new Date(),
+        logObject.fullPrefix.ToString,
+        errors
+      );
+    }
+  }
+  protected _handleLog<T extends any[]>(
     type: IOLevelLogId,
     message: IOStd<T>,
     color: typeof ForegroundColor = (chalk.Color = "white"),
     prefix?: string,
-    levelRange: IOLevelLog = 0
+    levelRange: IOLevelLog = 0,
+    errors: Error[] = []
   ): IOReturnType<T, P> {
     this.emit("willLog", type, message, prefix, levelRange, new Date());
-    const typeUpper = prefix ?? type.charAt(0).toUpperCase() + type.slice(1);
     const loggedAt = `${
       new Date().toLocaleTimeString() + " " + new Date().toLocaleDateString()
     }`;
-    const timeAndType: IOReturnGetTimeAndType = this.getTimeAndType(
-      typeUpper as "Log" | "Error" | "Info" | "Warn" | "Fatal" | "Debug",
+    const prefixName = prefix ?? type.charAt(0).toUpperCase() + type.slice(1);
+    const fullPrefix = Helper.getPrefix.apply(this, [
+      prefixName,
       color,
-      loggedAt
-    );
-    const stringToPrint = this.formatString(
-      `${this.short ? "" : `${message ? `${timeAndType.ToString}` : ""}`}`,
-      ...message
-    );
-    const ioLogObject: IOReturnType<T, P> = this.returnTypeFunction(
+      loggedAt,
+    ]);
+    const logObject: IOReturnType<T, P> = this._makeLogObject(
       type,
       message,
-      stringToPrint
+      fullPrefix,
+      prefixName,
+      levelRange,
+      color
     );
-    this.emit(
-      "logging",
-      this.levelLog,
-      ioLogObject,
-      this.formatString(...message),
-      stringToPrint,
-      new Date()
-    );
-    if (!this.checkLevel(levelRange)) return ioLogObject;
-    if (this.format === "pretty") {
-      this.write(stringToPrint);
-    } else if (this.format === "json") {
-      this.write(
-        `${this.short ? "" : `${message ? `${timeAndType.ToString}` : ""}`}`,
-        this.toJson(ioLogObject, this.censor(ioLogObject)),
-        "\n"
+    try {
+      this.emit(
+        "logging",
+        this._levelLog,
+        logObject,
+        this._formatString(...logObject.data),
+        new Date(),
+        logObject.fullPrefix.ToString
       );
+      if (!Helper.checkLevel.apply(this, [levelRange])) return logObject;
+      if (this._format === "pretty") {
+        this._printPrettyLog(logObject, errors);
+      } else if (this._format === "json") {
+        this._printJsonLog(logObject);
+      } else {
+        this._printHiddenLog(logObject, errors);
+      }
+    } catch (e) {
+      this.emit("error", e.stack as string, e as Error);
     }
-    return ioLogObject;
+    return logObject;
   }
-  protected getDataError<T extends object>(
-    errorList: IOErrorParam<T>,
-    detail: object = {}
-  ): IOReturnError[] {
-    const returnLogObject: IOReturnError[] = [];
-    for (const err of errorList.errors) {
-      const stack = parse(err);
-      returnLogObject.push({
-        nativeError: err.stack!,
-        detail,
-        user: this.hostname,
-        isError: true,
-        ...this.getErrorStack(stack),
-      });
-    }
-    return returnLogObject;
-  }
-  protected writePrettyFatal(err: Error) {
-    this.write(
+  protected _writePrettyFatal(err: Error) {
+    this._write(
       "",
       "Message:",
       chalk.redBright(err.message),
@@ -278,131 +290,76 @@ export class LoggerUtils<P extends {}> extends LoggerEvent {
       `--------------------------------------------------------------------------`
     );
   }
-  protected handleLogFatal<T extends object>(
-    errorList: IOErrorParam<T>
-  ): IOReturnType<IOReturnError[], P> {
-    this.emit("willLog", "fatal", errorList.errors, undefined, 1, new Date());
-    const loggedAt = `${
-      new Date().toLocaleTimeString() + " " + new Date().toLocaleDateString()
-    }`;
-    const timeAndType: IOReturnGetTimeAndType = this.getTimeAndType(
-      "Fatal",
-      (chalk.Color = "magentaBright"),
-      loggedAt
-    );
-    try {
-      const ioLogDataError: IOReturnError[] = this.getDataError(
-        errorList,
-        errorList.detail
-      );
-      let errorString = "";
-      const ioLogObject: IOReturnType<IOReturnError[], P> =
-        this.returnTypeFunction("fatal", ioLogDataError, "");
-      if (!this.checkLevel(1)) return ioLogObject;
-      if (errorList.errors.length < 1) return ioLogObject;
-      if (this.format === "pretty") {
-        this.write(
-          `${
-            this.short
-              ? ""
-              : `${
-                  errorList
-                    ? `${timeAndType.ToString}\n--------------------------------------------------------------------------`
-                    : ""
-                }`
-          }`
-        );
-        for (var i = 0, len = errorList.errors.length; i < len; i++) {
-          let err = errorList.errors[i];
-          this.writePrettyFatal(err);
-          errorString += this.formatString(err);
-        }
-        this.emit(
-          "logging",
-          this.levelLog,
-          ioLogObject,
-          errorString,
-          errorString,
-          new Date()
-        );
-        this.emit(
-          "fatalLogging",
-          this.levelLog,
-          ioLogObject,
-          new Date(),
-          timeAndType.ToString,
-          errorString,
-          ...errorList.errors
-        );
-        ioLogObject.ToString += errorString;
-      } else if (this.format === "json") {
-        this.write(
-          `${
-            this.short ? "" : `${errorList ? `${timeAndType.ToString}` : ""}`
-          }`,
-          this.toJson(ioLogObject, this.censor(ioLogObject))
-        );
-      }
-      return ioLogObject;
-    } catch (err: any) {
-      this.handleLogFatal({ errors: [err] });
-      return err;
-    }
-  }
-  protected returnTypeFunction<T extends any[]>(
+  protected _makeLogObject<T extends any[]>(
     type: IOLevelLogId,
     message: T,
-    ToString: string
+    fullPrefix: IOReturnGetTimeAndType,
+    prefix: string = "",
+    levelRange: IOLevelLog = 0,
+    color: typeof ForegroundColor = "black"
   ): IOReturnType<T, P> {
-    const stackObj = this.getErrorStack(undefined, 4);
+    const stackObj = Helper.getErrorStack(undefined, 4);
     return {
       levelLog: type,
       data: message,
-      loggedAt: this.getLoggedTime(),
-      hostName: this.hostname,
+      loggedAt: this._getLoggedTime(),
+      hostName: this._hostname,
       instanceName: this.loggerName,
-      cagetory: this.cagetoryName,
+      cagetory: this._cagetoryName,
       ...stackObj,
       setting: this.listSetting(),
-      bindingProps: this.childProps,
-      ToString: ToString,
+      bindingProps: this._childProps,
       toJson() {
         return JSON.stringify(this);
       },
-      pid: this.pid,
+      pid: this._pid,
+      prefix,
+      levelRange,
+      color,
+      fullPrefix,
     };
   }
   public warn<T extends any[]>(...message: IOStd<T>): IOReturnType<T, P> {
-    return this.handleLog(
+    return this._handleLog.apply(this, [
       "warn",
       message,
-      (chalk.Color = "yellowBright"),
+      "yellowBright",
       undefined,
-      2
-    );
+      2,
+    ]);
   }
   public error<T extends any[]>(...message: IOStd<T>): IOReturnType<T, P> {
-    return this.handleLog(
+    return this._handleLog.apply(this, [
       "error",
       message,
-      (chalk.Color = "redBright"),
+      "redBright",
       undefined,
-      1
-    );
+      1,
+    ]);
   }
   public info<T extends any[]>(...message: IOStd<T>): IOReturnType<T, P> {
-    return this.handleLog(
+    return this._handleLog.apply(this, [
       "info",
       message,
-      (chalk.Color = "cyanBright"),
+      "cyanBright",
       undefined,
-      4
-    );
+      4,
+    ]);
   }
   public fatal<T extends object>(
     error: IOErrorParam<T>
   ): IOReturnType<IOReturnError[], P> {
-    return this.handleLogFatal<T>(error);
+    const ioLogDataError: IOReturnError[] = Helper.getDataError.apply(this, [
+      error,
+      error.detail,
+    ]);
+    return this._handleLog.apply(this, [
+      "fatal",
+      ioLogDataError,
+      "magentaBright",
+      undefined,
+      1,
+    ]);
   }
   public prefix<T extends any[]>(
     opt: IOPrefixOption,
@@ -413,12 +370,12 @@ export class LoggerUtils<P extends {}> extends LoggerEvent {
       color: opt.color ?? "magentaBright",
       levelLog: opt.levelLog ?? 4,
     };
-    return this.handleLog(
+    return this._handleLog.apply(this, [
       "prefix",
       message,
       opt.color,
       opt.prefix,
-      opt.levelLog
-    );
+      opt.levelLog,
+    ]);
   }
 }