+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Start From 1.6.7 version
 
+## 1.6.11
+
+- Small fix, clean code, increase performance
+
+- Added colorized with `json` print type
+
 ## 1.6.10
 
 - Increase performance